docs(README): update balamod links to new repo
Balamod has changed its repo to be owned by an organization. This has affected the links for this repository.

According to the creator:

```
this project moved to https://github.com/balamod/balamod

Sorry if I'm not really active right now, it's a bit complex and I feel like it's slowing down the project.

To address this issue, I've decided to move balamod over to https://github.com/balamod/balamod in order to keep the project alive, and not be a bottleneck. I will also fork that repository into UwUDev/balamod to keep the mod menu working without having to patch balamod. Existing contributors to balamod will be invited to the balamod github organization in order to keep maintaining the project.
```